refactor(ai): estrai il prompt Smart Import
